+++ src/components/shared/navigationBars/topBar/topBar.css
@@ -139,3 +139,12 @@
     }
   }
 }
+
+.logoutBtn {
+  position: relative;
+  display: inline-block;
+  vertical-align: middle;
+  display: flex;
+  align-items: center;
+  padding-right: 24px;
+}
